+++ src/baby.js
@@ -3,19 +3,9 @@ import {TextField, RadioButtonGroup, RadioButton,DatePicker} from 'material-ui'
 import {Family as dbFamily} from './db'
 import RewardGoal from './components/rewards'
 
-const {List,CommandBar,Photo}=UI
+const {List,CommandBar,Photo, Messager}=UI
 
 export default class Baby extends Component{
-	constructor(){
-		super(...arguments)
-		this.state={}
-	}
-
-	componentWillReceiveProps(newProps){
-		if(this.props.params.name!=newProps.params.name)
-			dbFamily.currentChild=newProps.params.name
-	}
-
 	shouldComponentUpdate(newProps){
 		if(this.state.frozen)
 			return false
@@ -23,26 +13,13 @@ export default class Baby extends Component{
 		return this.props.child!=newProps.child
     }
 
+	componentWillReceiveProps(newProps){
+		if(this.props.params.name!=newProps.params.name)
+			dbFamily.currentChild=newProps.params.name
+	}
+
     render(){
         let {child}=this.props
-			,cmds=[]
-			,rewards
-
-        cmds.push(child._id ? "Remove" : "Save")
-
-		if(child._id){
-			rewards=(
-				<div>
-					<br/>
-						<br/>
-						<div style={{fontSize:"smaller", color:"gray", borderBottom:"1px dotted lightgray"}}>{child.name}的激励计划</div>
-						<RewardGoal
-							editable={true}
-							style={{marginTop:30}}
-							child={child}/>
-				</div>
-			)
-		}
         return (
             <div>
                 <div className="form">
@@ -50,15 +27,22 @@ export default class Baby extends Component{
                         <Photo ref="photo"
                             width={150}
                             height={150}
-                            src={child.photo} />
+                            src={child.photo}
+							onChange={url=>{
+								child.photo=url
+								dbFamily.upsert(child)
+							}}/>
                     </div>
-                    <TextField ref="name" floatingLabelText="child name"
+					
+                    <TextField ref="name" 
+						floatingLabelText="child name"
                         fullWidth={true}
                         value={child.name}
                         onBlur={e=>{
-							if(this.props.value!=e.target.value){
-								child.name=e.target.value
-								this.onChange(child)
+							let value=e.target.value.trim()
+							if(child.name!=value){
+								child.name=value
+								dbFamily.upsert(child)
 							}
 						}}/>
 
@@ -69,9 +53,9 @@ export default class Baby extends Component{
                         maxDate={new Date()}
                         value={child.bd}
                         onChange={(nil, date)=>{
-							if(this.props.value && this.props.value.getTime()!==date.getTime()){
+							if(!child.db || child.db.getTime()!=date.getTime()){
 								child.db=date
-								this.onChange(child)
+								dbFamily.upsert(child)
 							}
 						}}/>
 
@@ -79,81 +63,102 @@ export default class Baby extends Component{
                         style={{marginTop:36}}
                         name="gender"
                         onChange={(e,selected)=>{
-							if(this.props.valueSelected!=selected){
+							if(child.gender!=selected){
 								child.gender=selected
-								this.onChange(child)
+								dbFamily.upsert(child)
 							}
 						}}
                         valueSelected={child.gender||"f"}>
                         <RadioButton value="f" label="girl"/>
                         <RadioButton value="m" label="boy" />
                     </RadioButtonGroup>
-                    {rewards}
+					
+                    <div>
+						<br/>
+						<br/>
+						<div style={{fontSize:"smaller", color:"gray", borderBottom:"1px dotted lightgray"}}>
+							{child.name}的激励计划
+						</div>
+						<RewardGoal
+							editable={true}
+							style={{marginTop:30}}
+							child={child}/>
+					</div>
                 </div>
+				
                 <CommandBar className="footbar"
-                    items={cmds}
-                    onSelect={cmd=>this.onSelect(cmd)}
-                    />
+                    items={[{action:"Remove", onSelect:a=>this.remove()}]}/>
             </div>
         )
     }
-    onRewardRule(rule){
-        let {child}=this.props,
-            {rewardRules=[]}=child
-        rewardRules.push(rule)
-        child.rewardRules=rewardRules
-        this.onChange(child)
-    }
-
-    onReward(reward){
-        let {child}=this.props,
-            {rewardDetail=[]}=child
-        rewardDetail.push(reward)
-        child.rewardDetail=rewardDetail
-        this.onChange(child)
-    }
-
-    onChange(child){
-        if(!child._id)
-			return
-        dbFamily.upsert(child)
-			.then(a=>this.forceUpdate())
-    }
-    takePhoto(){
-        navigator.camera.getPicture((p)=>{
-                this.refs.photo.getDOMNode().src=p
-            }, (error)=>{
-
-            }, {
-                quality:50,
-                destinationType:Camera.DestinationType.FILE
-            });
-    }
-    onSelect(command){
-        var {child}=this.props
-        switch(command){
-        case "Save":
-            dbFamily.upsert(child)
-				.then(a=>this.context.router.replace(`baby/${child.name}`))
-            break
-        case "Remove":
-            this.setState({frozen:true})
-            dbFamily.remove(child._id)
-				.then(a=>this.context.router.replace("/"))
-            break
-        }
+	
+    remove(){
+        this.setState({frozen:true})
+		dbFamily.remove(this.props.child._id)
+			.then(a=>this.context.router.replace("/"))
     }
 
 	static contextTypes={router:React.PropTypes.object}
 
-	static Creator=class extends Baby{
-		componentWillUnmount(){
-			if(!this.props.child._id)
-				dbFamily.restoreLast()
-		}
-
+	static Creator=class extends Component{
 		componentWillReceiveProps(newProps){
 			return false
 		}
+		
+		render(){
+			return (
+				<div>
+					<div className="form">
+						<div className="child-photo">
+							<Photo ref="photo"
+								width={150}
+								height={150}/>
+						</div>
+						
+						<TextField ref="name" 
+							floatingLabelText="child name"
+							fullWidth={true}/>
+
+						<DatePicker ref="birthday" 
+							floatingLabelText="birthday"
+							fullWidth={true}
+							autoOk={true}
+							showYearSelector={true}
+							maxDate={new Date()}/>
+
+						<RadioButtonGroup ref="gender"
+							style={{marginTop:36}}
+							name="gender"
+							defaultSelected="f">
+							<RadioButton value="f" label="girl"/>
+							<RadioButton value="m" label="boy" />
+						</RadioButtonGroup>
+					</div>
+					
+					<CommandBar className="footbar"
+						items={[{action:"Save", onSelect:a=>this.save()}]}/>
+				</div>
+			)
+		}
+		
+		save(){
+			let {photo, name, birthday, gender}=this.refs
+			photo=photo.state.url
+			name=name.getValue()
+			birthday=birthday.getValue()
+			gender=gender.getValue()
+			
+			if(!name){
+				Messager.show("name can't be empty")
+				return
+			}
+			
+			Family.upsert({photo,name, gender, db:birthday})
+				.then(baby=>{
+					Family.currentChild=baby
+					this.context.router.replace(`baby/${name}`)
+				})
+		}
+		static contextTypes={router:React.PropTypes.object}
 	}
 }