+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 * remove moment dependency
 * create task from knowledge
+* elegant qiliApp: contextual, global back button, contextual object UI
 
 todos
 =====
 * replace calendar with material-ui.datepicker
+* Photo need getValue, and onChange(newUrl, oldUrl)
 
 lessons
 =======